+++ packages/transaction-pay-controller/src/types.ts
@@ -15,6 +15,7 @@ import type { GasFeeControllerActions } from '@metamask/gas-fee-controller';
 import type { Messenger } from '@metamask/messenger';
 import type { NetworkControllerFindNetworkClientIdByChainIdAction } from '@metamask/network-controller';
 import type { NetworkControllerGetNetworkClientByIdAction } from '@metamask/network-controller';
+import type { RampsControllerGetQuotesAction } from '@metamask/ramps-controller';
 import type { RemoteFeatureFlagControllerGetStateAction } from '@metamask/remote-feature-flag-controller';
 import type {
   AuthorizationList,
@@ -46,6 +47,7 @@ export type AllowedActions =
   | GasFeeControllerActions
   | NetworkControllerFindNetworkClientIdByChainIdAction
   | NetworkControllerGetNetworkClientByIdAction
+  | RampsControllerGetQuotesAction
   | RemoteFeatureFlagControllerGetStateAction
   | TokenBalancesControllerGetStateAction
   | TokenRatesControllerGetStateAction
@@ -396,6 +398,9 @@ export type TransactionPayFees = {
 
   /** Fee charged by the quote provider. */
   provider: FiatValue;
+
+  /** Fee charged by fiat on-ramp provider. */
+  fiatProvider?: FiatValue;
 
   /** Network fee for transactions on the source network. */
   sourceNetwork: {